fix: view 이동
