+++ src/syncMonitor/Main.java
@@ -1,14 +1,10 @@
 package syncMonitor;
 
-import syncMonitor.config.wrapper.DbConfig.OracleConfig;
-import syncMonitor.config.wrapper.DbConfig.TiberoConfig;
 import syncMonitor.config.YmlConfigWrapper;
 import syncMonitor.config.wrapper.DbConfig.TopologyConfig;
 import syncMonitor.config.wrapper.MonitorConfig;
 import syncMonitor.mode.Normal;
-import syncMonitor.mode.OnlyTibero;
-import view.View;
-import view.ViewOracleTibero;
+import syncMonitor.view.ViewOracleTibero;
 
 import java.util.List;
