+++ src/syncMonitor/mode/Normal.java
@@ -3,7 +3,7 @@ package syncMonitor.mode;
 import syncMonitor.config.wrapper.DbConfig.OracleConfig;
 import syncMonitor.config.wrapper.DbConfig.TiberoConfig;
 import syncMonitor.config.wrapper.DbConfig.TopologyConfig;
-import view.ViewOracleTibero;
+import syncMonitor.view.ViewOracleTibero;
 import syncMonitor.session.SyncMonitorSessionOracle;
 import syncMonitor.session.SyncMonitorSessionTibero;
