+++ src/syncMonitor/view/View.java
@@ -1,4 +1,4 @@
-package view;
+package syncMonitor.view;
 
 public interface View {
     void doPrint();
